+++ web/app/search/page.tsx
@@ -1,7 +1,9 @@
 "use client";
 
 import type React from "react";
-import { useState } from "react";
+import { useMemo, useState } from "react";
+import { useAll, useMatched, useMyID, usePendingFromMe } from "~/api/user";
+import FullScreenCircularProgress from "~/components/common/FullScreenCircularProgress";
 import Search from "~/components/search/search";
 import Table from "~/components/search/table";
 
@@ -15,12 +17,51 @@ export default function SearchPage({
 }) {
   const [query, setQuery] = useState<string>(searchParams?.query ?? "");
 
+  const {
+    state: { data },
+  } = useAll();
+  const {
+    state: { data: myId },
+  } = useMyID();
+  const initialData = useMemo(() => {
+    return data?.filter((item) => item.id !== myId && item.id !== 0) ?? null;
+  }, [data, myId]);
+
+  const {
+    state: { data: matches },
+  } = useMatched();
+
+  const {
+    state: { data: pending },
+  } = usePendingFromMe();
+
+  if (!initialData) {
+    return <FullScreenCircularProgress />;
+  }
+  // リクエストを送ってない人のみリクエスト送信可能
+  // FIXME: O(n^2) | n = count(users) なのでめっちゃ計算コストかかる。なんとかして。
+  const canRequest = (userId: number) =>
+    !matches?.some((match) => match.id === userId) &&
+    !pending?.some((pending) => pending.id === userId);
+
+  // this is very expensive. someone fix this pls
+  const users = initialData?.filter(
+    (user) =>
+      query === "" ||
+      user.name.includes(query) ||
+      user.interestSubjects.some((i) => i.name.includes(query)),
+  );
+
   return (
     <div className="flex min-h-screen justify-center ">
       <div className="w-full">
         <h2 className="m-5 mb-4 font-bold text-2xl">ユーザー検索</h2>
         <Search placeholder="検索" setSearchString={setQuery} />
-        <Table query={query} />
+        {users ? (
+          <Table users={users} canRequest={canRequest} />
+        ) : (
+          <span>ユーザーが見つかりません</span>
+        )}
       </div>
     </div>
   );
